Add files via upload
The datasets employed in our study were sourced from prominent LLPS databases such as DrLLPS and PhaSePro.. One notable observation during the training process was the prevalence of repetitive amino acids in natural proteins. This inherent characteristic tends to get accentuated in the generated proteins, resulting in sequences with high repetitiveness. We implemented a repetitiveness score to address this issue, allowing us to filter out sequences that exhibited excessive repetitiveness.